subarea analysis
Given subarea.csv defined from NeXTA by users. DTALite first identifies the zones inside the subarea, then we select another 50% outside zones with highest volume related to the subarea as the origin zones for the subarea analysis. The reduced origin zone size (typically 10-20% of zones in the network) will help to generate the results rapidly. Users can further provide the traffic counts for the second stage of ODME and perform the sensitivity analysis and simulation.